requirements.txt
click>=8.1
pyyaml>=6.0
requests>=2.28
pytest>=7.0
